requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.9
pydotenv==0.0.7
PyJWT==2.1.0
PyMySQL==1.0.2
python-dotenv==0.19.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
